requirements.txt
################################################################################
# This requirements file has been automatically generated from `Pipfile` with
# `pipenv-to-requirements`
#
#
# This has been done to maintain backward compatibility with tools and services
# that do not support `Pipfile` yet.
#
# Do NOT edit it directly, use `pipenv install [-d]` to modify `Pipfile` and
# `Pipfile.lock` and then regenerate `requirements*.txt`.
################################################################################

asgiref==3.5.0 ; python_version >= '3.7'
async-timeout==4.0.2 ; python_version >= '3.6'
attrs==21.4.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
babel==2.9.1 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
bcrypt==3.2.0
blinker==1.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12 ; python_version >= '3'
click==8.1.2 ; python_version >= '3.7'
decorator==5.1.1 ; python_version >= '3.5'
deprecated==1.2.13 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
distlib==0.3.4
django
dnspython==2.2.1 ; python_version >= '3.6' and python_version < '4.0'
email-validator==1.1.3 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
filelock==3.6.0 ; python_version >= '3.7'
flake8-docstrings==1.6.0
flake8==4.0.1
flask-babelex==0.9.4
flask-login==0.6.0
flask-mail==0.9.1
flask-mongoengine==1.0.0
flask-principal==0.4.0
flask-rbac==0.5.0
flask-recaptcha==0.4.2
flask-redis==0.4.0
flask-security==3.0.0
flask-sqlalchemy==2.5.1
flask-wtf==1.0.1
flask==2.1.1
h11==0.13.0 ; python_version >= '3.6'
idna==3.3 ; python_version >= '3'
infinity==1.5
iniconfig==1.1.1
intervals==0.9.2
itsdangerous==2.1.2 ; python_version >= '3.7'
jinja2==3.1.1 ; python_version >= '3.7'
markupsafe==2.0.1
mccabe==0.6.1
mongoengine==0.24.1
mongopy==0.1
nose==1.3.7
packaging==21.3 ; python_version >= '3.6'
passlib==1.7.4
pbr==5.8.1 ; python_version >= '2.6'
pendulum==2.1.2
pip==22.0.4 ; python_version >= '3.7'
pipenv-to-requirements==0.9.0
pipenv==2022.3.28
platformdirs==2.5.1 ; python_version >= '3.7'
pluggy==1.0.0 ; python_version >= '3.6'
postgres==4.0
psycopg2-binary==2.9.3 ; python_version >= '3.6'
psycopg2-pool==1.1 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
psycopg2==2.9.3
py==1.11.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
pycodestyle==2.8.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
pycparser==2.21
pydocstyle==6.1.1 ; python_version >= '3.6'
pyflakes==2.4.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pymongo==4.1.1 ; python_full_version >= '3.6.2'
pyparsing==3.0.8 ; python_full_version >= '3.6.8'
pytest==7.1.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
pytzdata==2020.1 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
redis==4.2.2 ; python_version >= '3.6'
requests==2.27.1
setuptools==62.1.0 ; python_version >= '3.7'
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
snowballstemmer==2.2.0
speaklater==1.3
sqlalchemy-utils==0.38.2 ; python_version ~= '3.4'
sqlalchemy==1.4.34
sqlparse==0.4.2 ; python_version >= '3.5'
tomli==2.0.1 ; python_version >= '3.7'
urllib3==1.26.9 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4.0'
validators==0.18.2 ; python_version >= '3.4'
virtualenv-clone==0.5.7 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
virtualenv==20.14.1 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
werkzeug==2.1.1
wrapt==1.14.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
wsproto==1.1.0
wtforms-alchemy==0.18.0
wtforms-components==0.10.5
wtforms==3.0.1
